requirements.txt
openai==0.27.0
PyPDF2==3.0.1
python-dotenv==1.0.0
